Crate libexif

The exif crate provides a safe wrapper around the libexif C library. It provides the ability to read EXIF data from image files. The entry point for inspecting a file’s EXIF data is Data::open(). EXIF data can be inspected by iterating over the data’s contents and entries:

fn dump_exif<P: AsRef<Path>>(file_name: P) -> Result<(), libexif::ExifError> {
    let data = libexif::Data::open("image.jpg")?;

    for content in data.contents() {
        println!("[{:=>32}{:=>46}]", format!(" {:?} ", content.ifd()), "");

        for entry in content.entries() {
            println!("  {:<30} = {}",
                     entry.tag().title(content.ifd()),
                     entry.text_value()?);
        }
    }

    Ok(())
}

Re-exports

pub use error::*;

Modules

error

Structs

Content

Container for all EXIF data in a single IFD.

Data

Container for all EXIF data found in an image.

Entry

Data found in a single EXIF tag.

Rational

A rational number consisting of a numerator and denominator.

Tag

EXIF tag.

Enums

ByteOrder

Defines the byte order of binary values.

DataEncoding

Defines the encoding used to represent EXIF data.

DataOption

Options that affect the behavior of Data.

DataType

EXIF tag data formats.

ExifIfd

Possible EXIF Image File Directories

ExifTag

EXIF tags

IFD

Image file directory types.

SupportLevel

Requirement specificatoins for standard EXIF tags.

Value

Dynamic value for an EXIF tag.

Constants

EXIF_HEADER

JPEG_HEADER

Type Definitions

ExifEntry

Data found in one EXIF tag