Enum libexif::Value

pub enum Value {
    Text(String),
    U8(Vec<u8>),
    I8(Vec<i8>),
    U16(Vec<u16>),
    I16(Vec<i16>),
    U32(Vec<u32>),
    I32(Vec<i32>),
    URational(Vec<Rational<u32>>),
    IRational(Vec<Rational<i32>>),
    Undefined(Vec<u8>),
}

Dynamic value for an EXIF tag.

Each variant of Value corresponds to a variant of DataType. Each variant (except for Text) contains a Vec with a length equal to the number of components in the corresponding Entry.

Variants

Text(String)

Tuple Fields

0: String

Value interpreted as a string.

U8(Vec<u8>)

Tuple Fields

0: Vec<u8>

Value interpreted as unsigned bytes.

I8(Vec<i8>)

Tuple Fields

0: Vec<i8>

Value interpreted as signed bytes.

U16(Vec<u16>)

Tuple Fields

0: Vec<u16>

Value interpreted as unsigned 16-bit integers.

I16(Vec<i16>)

Tuple Fields

0: Vec<i16>

Value interpreted as signed 16-bit integers.

U32(Vec<u32>)

Tuple Fields

0: Vec<u32>

Value interpreted as unsigned 32-bit integers.

I32(Vec<i32>)

Tuple Fields

0: Vec<i32>

Value interpreted as signed 32-bit integers.

URational(Vec<Rational<u32>>)

Tuple Fields

0: Vec<Rational<u32>>

Value interpreted as unsigned Rational numbers.

IRational(Vec<Rational<i32>>)

Tuple Fields

0: Vec<Rational<i32>>

Value interpreted as signed Rational numbers.

Undefined(Vec<u8>)

Tuple Fields

0: Vec<u8>

Value is uninterpreted sequence of bytes.

Implementations

impl Value

pub fn unwrap_u8(&self) -> Vec<u8>

pub fn unwrap_i8(&self) -> Vec<i8>

pub fn unwrap_u16(&self) -> Vec<u16>

pub fn unwrap_i16(&self) -> Vec<i16>

pub fn unwrap_u32(&self) -> Vec<u32>

pub fn unwrap_i32(&self) -> Vec<i32>

pub fn unwrap_undefined(&self) -> Vec<u8>

pub fn unwrap_text(&self) -> String

pub fn get_components_size_format(
&self
) -> Result<(usize, usize, ExifFormat), ExifError>