Struct libexif::Data

pub struct Data { /* fields omitted */ }

Container for all EXIF data found in an image.

Implementations

impl Data

pub fn new() -> Self

Create an empty EXIF data

pub fn open<P: AsRef<Path>>(path: P) -> Result<Data, ExifError>

Construct a new EXIF data container with EXIF data from a JPEG file.

pub fn byte_order(&self) -> ByteOrder

Return the byte order in use by this EXIF data.

pub fn set_byte_order(&mut self, byte_order: ByteOrder)

Set the byte order used for this EXIF data.

pub fn encoding(&self) -> DataEncoding

Return the encoding in use by this EXIF data.

pub fn set_encoding(&mut self, encoding: DataEncoding)

Set the encoding used for this EXIF data.

pub fn set_option(&mut self, option: DataOption)

Enable a data processing option.

pub fn unset_option(&mut self, option: DataOption)

Disable a data processing option.

pub fn get_entry(
&self,
ifd: impl ToLibExif<ExifIfd>,
tag: ExifTag
) -> Result<Entry<'_>, ExifError>

Get a Entry

Example

    use libexif::{IFD, ExifTag, ByteOrder, Data};
    let mut data = Data::open("assets/file.jpg").unwrap();
    let _ = data
        .get_entry(IFD::Image, ExifTag::EXIF_TAG_X_RESOLUTION)
        .unwrap()
        .value(ByteOrder::LittleEndian);

pub fn get_entry_raw(
&self,
ifd: impl ToLibExif<ExifIfd>,
tag: ExifTag
) -> Result<&mut ExifEntry, ExifError>

Returns a raw entry directly from the C bindings without any wrapper Same as get_entry().inner

Saftey

This method dereferences a raw pointer

pub fn set_entry(
&mut self,
ifd: IFD,
tag: ExifTag,
value: Value,
order: ByteOrder
) -> Result<(), ExifError>

Set an entry

Example

use libexif::{IFD, ExifTag, ByteOrder, Data, Value};
let mut data = Data::open("assets/file.jpg").unwrap();
data.set_entry(
    IFD::Image,
    ExifTag::EXIF_TAG_ORIENTATION,
    Value::U16(vec![100u16]),
    ByteOrder::LittleEndian,
);

Need to know IFD, Tag, Value and ByteOrder Before Hand

pub fn contents(&self) -> impl ExactSizeIterator<Item = Content<'_>>

Iterate over the contents of the EXIF data.

pub fn raw_data(&self) -> &[u8]

Return the raw binary data for the ExifData

pub fn fix(&mut self)

Fix the EXIF data to make it compatible with the EXIF specification.

pub fn dump(&self)

Dump all EXIF data to stdout.

pub fn write_from_file(
&mut self,
from: impl AsRef<Path>,
to: impl AsRef<Path>
) -> Result<(), ExifError>

pub fn write(
&mut self,
old_buffer: impl AsRef<[u8]>,
to: impl AsRef<Path>
) -> Result<(), ExifError>